moved FloatToQuantizedStruct and QuantizedToFloatStruct to quantization_utils.hpp
